+++ components/LeaguePage/StaffView.tsx
@@ -1,11 +1,11 @@
 import React, { useContext } from "react";
-import styles from "../styles/components/LeagueViews.module.css";
+import styles from "../../styles/components/LeagueViews.module.css";
 import { StaffCard } from "./StaffCard";
-import { APIContext } from "../context/APIContext";
-import { Loader } from "./util/Loader";
-import { Error } from "./util/Error";
-import { Empty } from "./util/Empty";
-import { User } from "../models/users";
+import { APIContext } from "../../context/APIContext";
+import { Loader } from "../util/Loader";
+import { Error } from "../util/Error";
+import { Empty } from "../util/Empty";
+import { User } from "../../models/users";
 import useSWR from "swr";
 
 const filters = [
