+++ components/LeaguePage/StudentCard.tsx
@@ -1,5 +1,5 @@
 import React from "react";
-import styles from "../styles/components/LeagueViews.module.css";
+import styles from "../../styles/components/LeagueViews.module.css";
 
 type StudentCardProps = {
   firstName: string;
