+++ pages/league.tsx
@@ -2,9 +2,9 @@ import { NextPage } from "next";
 import React, { useState, useEffect, useContext } from "react";
 import { APIContext } from "../context/APIContext";
 import styles from "../styles/League.module.css";
-import { ClassView } from "../components/ClassView";
-import { StudentView } from "../components/StudentView";
-import { StaffView } from "../components/StaffView";
+import { ClassView } from "../components/LeaguePage/ClassView";
+import { StudentView } from "../components/LeaguePage/StudentView";
+import { StaffView } from "../components/LeaguePage/StaffView";
 import { Class } from "../models/class";
 import { User } from "../models/users";
 import { Student } from "../models/students";
